Properly wrapping graph drawing in own React component.
