+++ pydash-front/src/app/overview/DashTile.js
@@ -1,9 +1,13 @@
 import React, {Component} from 'react';
+import axios from 'axios';
+
 import './overview.css';
 import Grid from 'material-ui/Grid';
 import Card, { CardContent } from 'material-ui/Card';
 import { withStyles } from 'material-ui/styles';
 import Typography from 'material-ui/Typography';
+
+import DashboardVisitsGraph from './DashboardVisitsGraph'
 
 const styles = {
     card: {
@@ -13,23 +17,68 @@ const styles = {
     }
 };
 
+// Transforms a hashmap of key-value pairs into an array of {x: key, y: value} objects.
+// TODO move to a helper JS file.
+function dict_to_xy_arr(dict){
+    let res =  Object.entries(dict).map(function([key, value]){
+        return {x: key, y: value}
+    });
+    console.log('dict_to_xy_array', res);
+    return res;
+}
+
+
 class DashTile extends Component {
     constructor(props) {
         super(props);
         this.state = {
-            title: props.title,
             col: props.xs,
+            total_visits: "?",
+            visits_per_day: [],
         };
     }
-    
+
+    componentDidMount() {
+        console.log("TEST")
+        // TODO: Get dashboard_id from newProps
+        const dashboard_id = this.props.dashboard_id
+        axios('http://localhost:5000/api/dashboards/' + dashboard_id, {
+            method: 'get',
+            withCredentials: true
+        }).then((response) => {
+            console.log('success', response);
+            this.setState(prevState =>{
+                let newState = prevState;
+                newState.total_visits = "" + response.data.aggregates.total_visits;
+                newState.visits_per_day = dict_to_xy_arr(response.data.aggregates.visits_per_day)
+                console.log(newState)
+
+                return newState;
+            })
+        }).catch((error) => {
+            console.log('error', error);
+
+            /* this.setState(prevState =>{
+                *     let newState = prevState;
+                *     newState.total_visits = "" + mock_data.aggregates.total_visits;
+                *     newState.visits_per_day = dict_to_xy_arr(mock_data.aggregates.visits_per_day)
+                *     console.log(newState)
+
+                *     return newState;
+                * })
+            */
+        });
+    }
+
+
     render() {
         return(
             <Grid item xs={this.state.col}>
                 <Card>
                     <CardContent>
-                        <Typography variant='headline' component='h2'>
-                            {this.state.title}
-                        </Typography>
+                        <Typography component='h2'>{this.props.title}</Typography>
+                        <Typography>Total Visits: {this.state.total_visits}</Typography>
+                        <DashboardVisitsGraph data={this.state.visits_per_day} title="Visits Per Day" tooltip_title="No. visits: "/>
                     </CardContent>
                 </Card>
             </Grid>
